+++ NEWS.md
@@ -2,6 +2,8 @@
    
 ## New features
    
+- New helper function `ggcumevents()` for ploting the cumulative number of events table ([#117](https://github.com/kassambara/survminer/issues/117)).
+
 - New function `+.ggsurv()` to add ggplot components - `theme()`, `labs()` -  to an object of class ggsurv, which is a list of ggplots. ([#151](https://github.com/kassambara/survminer/issues/151)). For example:
 
 ```r